+++ tomcat/src/main/java/org/apache/coyote/exception/HttpConnectorException.java
@@ -1,0 +1,14 @@
+package org.apache.coyote.exception;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public abstract class HttpConnectorException extends RuntimeException {
+
+    private static final Logger log = LoggerFactory.getLogger(HttpConnectorException.class);
+
+    HttpConnectorException(String message) {
+        super(message);
+        log.error(message);
+    }
+}
