+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequestReceiver.java
@@ -6,7 +6,7 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
-import org.apache.coyote.http11.exception.ReceivingRequestFailedException;
+import org.apache.coyote.exception.ReceivingRequestFailedException;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.HttpRequestBody;
 import org.apache.coyote.http11.request.HttpRequestHeader;
