--- tomcat/src/main/java/org/apache/coyote/http11/exception/HttpConnectorException.java
@@ -1,8 +1,0 @@
-package org.apache.coyote.http11.exception;
-
-public abstract class HttpConnectorException extends RuntimeException {
-
-     HttpConnectorException(String message) {
-        super(message);
-    }
-}
